requirements.txt
deep-translator==1.11.4
einops==0.8.0
groq==0.8.0
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llama-index-llms-ollama==0.1.4
llamaindex-py-client==0.1.19
ollama==0.1.8
openai==1.30.1
openai-whisper==20231117
PyAudio==0.2.14
pyttsx3==2.90
regex==2024.4.16
requests==2.32.3
requests-oauthlib==2.0.0
threadpoolctl==3.4.0
tkhtmlview==0.3.0
tokenizers==0.19.1
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.10.0
vosk==0.3.45
wasm==1.2
